+++ src/Enum/ZplMode.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\Enum;
+
+enum ZplMode: string
+{
+    case Zpl = '1';
+    case ZplII = '2';
+}